Added PlayerStrategy attribute to Player, added Player constructor, getters and setters for PlayerStrategy, updated initGameDummy method, added Human decision logic
