+++ GameEngine/GameEngineDriver.cpp
@@ -19,14 +19,14 @@ int gameEngineDriver(GAME_MODE gm, const string &fileArg) {
     PHASE phase;
 
     // startup
-    phase = START;
-    game->startupPhase(game, cp, command, phase, observer);
+    // phase = START;
+    // game->startupPhase(game, cp, command, phase, observer);
 
     // main game loop
-//    phase = CHECK_WIN;
-//    game->initGameDummy(observer);
+    phase = CHECK_WIN;
+    game->initGameDummy(observer);
 //    game->initGameEndDummy(observer);
-//    game->mainGameLoop(game, phase, observer);
+    game->mainGameLoop(game, phase, observer);
 
 
 }
